+++ lib/onfido/errors/connection_error.rb
@@ -1,4 +1,4 @@
 module Onfido
-  class ConnectionError < StandardError
+  class ConnectionError < OnfidoError
   end
 end