+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add base errors class (`Onfido::OnfidoError`)
 - Add rubocop, and fix style inconsistencies
 
 ## v0.0.4 10 November 2015